+++ Altinn.AccessManagement.sln
@@ -15,6 +15,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.AccessManagement.Per
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.Authorization.ServiceDefaults", "..\altinn-authorization-utils\src\Altinn.Authorization.ServiceDefaults\src\Altinn.Authorization.ServiceDefaults\Altinn.Authorization.ServiceDefaults.csproj", "{6D847BD9-8BD8-44C1-810B-AC44D4DC0499}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.Platform.Authorization", "..\altinn-authorization\src\Authorization\Altinn.Platform.Authorization.csproj", "{D9C0433E-738C-4E07-9F88-D71D7DF04E3F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -45,6 +47,10 @@ Global
 		{6D847BD9-8BD8-44C1-810B-AC44D4DC0499}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6D847BD9-8BD8-44C1-810B-AC44D4DC0499}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6D847BD9-8BD8-44C1-810B-AC44D4DC0499}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D9C0433E-738C-4E07-9F88-D71D7DF04E3F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D9C0433E-738C-4E07-9F88-D71D7DF04E3F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D9C0433E-738C-4E07-9F88-D71D7DF04E3F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D9C0433E-738C-4E07-9F88-D71D7DF04E3F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
